Extract http client functionalities
